app/requirements.txt
Django==3.1.7
gunicorn==20.1.0
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
six==1.15.0

